+++ CHANGELOG.md
@@ -28,3 +28,4 @@
 
 - Added account endpoint
 - Added account program endpoint
+- Added account region endpoint